Uploaded bus file
This program asks the user to input how many passengers they have seen get on buses in a certain time frame. This program uses a while loop which is only exited when the user inputs a special character. At the end of the program the number of passengers and buses it's printed.